+++ CHANGELOG.md
@@ -24,6 +24,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - **Breaking**: The `size` field of the `nlist` struct is now a `size_t`.
 - **Breaking**: The `n_nodes` field of the `Agraphinfo_t` struct is now a
   `size_t`.
+- **Breaking**: The `nspans` field of `textlabel_t.u.txt` is now a `size_t`.
 
 ### Fixed